Rearrange shebang and constants
