+++ src/utils.ts
@@ -1,5 +1,10 @@
+import * as fs from 'fs';
 import * as net from 'net';
+import * as path from 'path';
+import { ENV_BIN, UTF8 } from './constants';
 import * as logger from './logger';
+
+const MATCHES_SHEBANG = /#!( *\S+ +)?( *\S+ *)$/;
 
 export interface PortError extends Error {
   message: string;
@@ -65,4 +70,46 @@ export const getAvailablePort = (callback: ServerCallback, attempt = 0) => {
         .close();
     })
     .listen(0);
+};
+
+export const handleShebang = (command: string): string => {
+  const filePath = path.join(process.cwd(), command);
+  const envFilePath = path.join(process.cwd(), ENV_BIN, command);
+
+  // In root of project
+  if (!fs.existsSync(filePath)) {
+    // In env
+    if (fs.existsSync(envFilePath)) {
+      return envFilePath;
+    // Global command
+    } else {
+      return command;
+    }
+  }
+
+  const firstLine = fs.readFileSync(filePath, UTF8).split('\n')[0];
+
+  if (!firstLine) {
+    return command;
+  }
+
+  const shebang = MATCHES_SHEBANG.exec(firstLine);
+
+  if (shebang) {
+    const shebangCommand = shebang[2].trim();
+
+    const shebangCommandPath = path.join(process.cwd(), ENV_BIN, shebangCommand);
+
+    const envExists = fs.existsSync(shebangCommandPath);
+
+    // Shebang in env
+    if (envExists) {
+      return `${shebangCommandPath} ${command}`;
+    }
+
+    // Global command
+    return `${shebangCommand} ${command}`;
+  }
+
+  return command;
 };
